requirements.txt
asgiref==3.4.1
click==8.0.1
Django==3.2.6
django-rest-framework==0.1.0
djangorestframework==3.12.4
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pytz==2021.1
sqlparse==0.4.1
Werkzeug==2.0.1
